Moving field classes, edit notebook import
